fix adamw

+++ proto/include/diopi/functions.h
@@ -2829,7 +2829,7 @@ DIOPI_API diopiError_t diopiReciprocalInp(diopiContextHandle_t ctx, diopiTensorH
  */
 DIOPI_API diopiError_t diopiFusedAdamW(diopiContextHandle_t ctx, diopiTensorHandle_t* params, diopiConstTensorHandle_t* grads, diopiTensorHandle_t* exp_avgs,
                                        diopiTensorHandle_t* exp_avg_sqs, diopiTensorHandle_t* max_exp_avg_sqs, diopiConstTensorHandle_t* state_steps,
-                                       int64_t nums, float lr, float beta1, float beta2, float eps, float weight_decay, bool amsgrad, bool maximize);
+                                       int64_t nums, double lr, double beta1, double beta2, double eps, double weight_decay, bool amsgrad, bool maximize);
 
 /**
  * @brief The function is used to implement the AdamW optimizer. Its functionality is to perform a single parameter update.
